requirements.txt
# Core Framework
streamlit>=1.30.0

# LLM & AI
langchain>=0.1.0
langchain-community>=0.0.1
openai>=1.0.0

# Vector Store & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# PDF Processing
pypdf>=3.0.0

# Database
supabase>=2.0.0

# Data Validation
pydantic>=2.0.0
email-validator>=2.0.0

# Date/Time Parsing
python-dateutil>=2.8.0

# Utilities
python-dotenv>=1.0.0
